Introducing CoWOptimization a library that uses macro to implement CoW
### Motivation

In order to implement the CoW of a struct, we need to define a
new storage class and cover all the setters and getters of the
struct, which may be a heavy job.

This macro helps us to quickly realize the structure of cows and
reduce a lot of repetitive work.

### Modifications

- Add CoWOptimization.
- Add CoWOptimizationMacro.
- Add tests for CoWOptimizationMacro.

### Result

CoWOptimization is ready for use.